+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Visit the package at [pub.dev](https://pub.dev/packages/solidpod).
 
 + Update permission table to list [0.9.18 20260129 jesscmoore]
 + Fixes background of loadingScreen [0.9.17 20260128 jesscmoore]
++ Integrate webid dialogs into grant permission form [0.9.16 20260128 jesscmoore]
 + Optimise shared individual key loading and retrieval [0.9.15 20260128 cdawei]
 + Restored retrieve permissions button [0.9.14 20260125 jesscmoore]
 + Export ResourceNotExistException [0.9.13 20260123 tonypioneer]